--- src/main/java/org/jboss/as/websockets/Frame.java
@@ -1,8 +1,0 @@
-package org.jboss.as.websockets;
-
-/**
- * @author Mike Brock
- */
-public interface Frame {
-  public FrameType getType();
-}
