--- src/main/java/org/jboss/as/websockets/FrameType.java
@@ -1,10 +1,0 @@
-package org.jboss.as.websockets;
-
-import java.io.InputStream;
-
-/**
- * @author Mike Brock
- */
-public enum FrameType {
-  Continuation, Text, Binary, Ping, Pong, ConnectionClose, Unknown
-}
