+++ src/main/java/org/jboss/as/websockets/WebSocket.java
@@ -16,9 +16,10 @@
 
 package org.jboss.as.websockets;
 
+import org.jboss.websockets.oio.OioWebSocket;
+
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpSession;
-import java.io.IOException;
 
 /**
  * Represents a handle to a single WebSocket connection. It has reader and writer methods to get data in and out.
@@ -27,36 +28,11 @@ import java.io.IOException;
  *
  * @author Mike Brock
  */
-public interface WebSocket {
-
-  /**
-   * A unique ID associated with the socket, which can be used for session association. This ID is generated by the
-   * WebSockets framework as a random hash when the socket is open and has no association with any external API or
-   * the websocket handshake process.
-   *
-   * @return A hex string representing the unique ID of the socket.
-   */
-  public String getSocketID();
-
-  /**
-   * Read a single frame from the socket.
-   *
-   * @return
-   * @throws IOException
-   */
-  public Frame readFrame() throws IOException;
+public interface WebSocket extends OioWebSocket
+{
 
 
-  /**
-   * Write a frame to the socket.
-   *
-   * @param frame
-   * @throws IOException
-   */
-  public void writeFrame(Frame frame) throws IOException;
-
-
-  /**
+   /**
    * Return the HTTP Session with which this WebSocket is associated.
    * @return an instance of the HttpSession
    */
@@ -72,8 +48,4 @@ public interface WebSocket {
    */
   public HttpServletRequest getServletRequest();
 
-  /**
-   * Terminates the connection with the client and closes the socket.
-   */
-  public void closeSocket() throws IOException;
 }
