--- src/main/java/org/jboss/as/websockets/frame/PingFrame.java
@@ -1,12 +1,0 @@
-package org.jboss.as.websockets.frame;
-
-import org.jboss.as.websockets.FrameType;
-
-/**
- * @author Mike Brock
- */
-public class PingFrame extends AbstractFrame {
-  public PingFrame() {
-    super(FrameType.Ping);
-  }
-}
