--- src/main/java/org/jboss/as/websockets/frame/PongFrame.java
@@ -1,12 +1,0 @@
-package org.jboss.as.websockets.frame;
-
-import org.jboss.as.websockets.FrameType;
-
-/**
- * @author Mike Brock
- */
-public class PongFrame extends AbstractFrame {
-  public PongFrame() {
-    super(FrameType.Pong);
-  }
-}
