+++ src/main/java/org/jboss/websockets/Frame.java
@@ -1,0 +1,8 @@
+package org.jboss.websockets;
+
+/**
+ * @author Mike Brock
+ */
+public interface Frame {
+  public FrameType getType();
+}
