+++ src/main/java/org/jboss/websockets/FrameType.java
@@ -1,0 +1,10 @@
+package org.jboss.websockets;
+
+import java.io.InputStream;
+
+/**
+ * @author Mike Brock
+ */
+public enum FrameType {
+  Continuation, Text, Binary, Ping, Pong, ConnectionClose, Unknown
+}
