+++ src/main/java/org/jboss/websockets/frame/CloseFrame.java
@@ -1,0 +1,12 @@
+package org.jboss.websockets.frame;
+
+import org.jboss.websockets.FrameType;
+
+/**
+ * @author Mike Brock
+ */
+public class CloseFrame extends AbstractFrame {
+  public CloseFrame() {
+    super(FrameType.ConnectionClose);
+  }
+}
