+++ src/main/java/org/jboss/websockets/frame/PingFrame.java
@@ -1,0 +1,12 @@
+package org.jboss.websockets.frame;
+
+import org.jboss.websockets.FrameType;
+
+/**
+ * @author Mike Brock
+ */
+public class PingFrame extends AbstractFrame {
+  public PingFrame() {
+    super(FrameType.Ping);
+  }
+}
