+++ src/main/java/org/jboss/websockets/oio/internal/protocol/ietf08/Hybi08Handshake.java
@@ -1,0 +1,30 @@
+/*
+ * Copyright 2012 JBoss, by Red Hat, Inc
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.jboss.websockets.oio.internal.protocol.ietf08;
+
+import org.jboss.websockets.oio.internal.protocol.ietf07.Hybi07Handshake;
+
+/**
+ * The handshaking protocol impelemtation for Hybi-07, which is identical to Hybi-08, and thus is just a thin
+ * subclass of {@link Hybi07Handshake} that sets a different version number.
+ * @author Mike Brock
+ */
+public class Hybi08Handshake extends Hybi07Handshake {
+  public Hybi08Handshake() {
+    super("8");
+  }
+}
